refactor(health-sdk): Extracted common methods for bottoms sheets into an inheritable class
IPC-245
